Finished renaming files. All working with new names.
